fixed up EpiStudy classes so that they import
