+++ banana/study/mri/epi.py
@@ -1,10 +1,8 @@
 from .base import MriStudy
-from nipype.interfaces.fsl import TOPUP, ApplyTOPUP
+from nipype.interfaces.fsl import (
+    TOPUP, ApplyTOPUP, Merge as FslMerge, BET, FUGUE, PrepareFieldmap)
 from nipype.interfaces import fsl
 from nipype.interfaces.utility import Merge as merge_lists
-from nipype.interfaces.fsl.utils import Merge as fsl_merge
-from nipype.interfaces.fsl.epi import PrepareFieldmap
-from nipype.interfaces.fsl.preprocess import BET, FUGUE
 from arcana.study import ParamSpec, SwitchSpec
 from arcana.data import InputFilesetSpec, FilesetSpec, FieldSpec
 from arcana.study.base import StudyMetaClass
@@ -27,22 +25,12 @@ from banana.requirement import mrtrix_req
 class EpiStudy(MriStudy, metaclass=StudyMetaClass):
 
     add_data_specs = [
-        InputFilesetSpec('series', STD_IMAGE_FORMATS,
-                         desc=("The set of EPI volumes that make up the "
-                               "series")),
         InputFilesetSpec('coreg_ref_wmseg', STD_IMAGE_FORMATS,
                          optional=True),
-        InputFilesetSpec('reverse_phase', STD_IMAGE_FORMATS, optional=True),
         InputFilesetSpec('field_map_mag', STD_IMAGE_FORMATS,
                          optional=True),
         InputFilesetSpec('field_map_phase', STD_IMAGE_FORMATS,
                          optional=True),
-        FilesetSpec('moco', nifti_gz_format,
-                    'intrascan_alignment_pipeline'),
-        FilesetSpec('align_mats', motion_mats_format,
-                    'intrascan_alignment_pipeline'),
-        FilesetSpec('moco_par', par_format,
-                    'intrascan_alignment_pipeline'),
         FieldSpec('field_map_delta_te', float,
                   'field_map_time_info_pipeline')]
 
@@ -59,96 +47,8 @@ class EpiStudy(MriStudy, metaclass=StudyMetaClass):
         if ('field_map_phase' in self.input_names and
                 'field_map_mag' in self.input_names):
             return self._fugue_pipeline(**name_maps)
-        elif 'reverse_phase' in self.input_names:
-            return self._topup_pipeline(**name_maps)
         else:
             return super().prepare_pipeline(**name_maps)
-
-    def _topup_pipeline(self, **name_maps):
-
-        pipeline = self.new_pipeline(
-            name='preprocess_pipeline',
-            desc=("Topup distortion correction pipeline"),
-            citations=[fsl_cite],
-            name_maps=name_maps)
-
-        reorient_epi_in = pipeline.add(
-            'reorient_epi_in',
-            fsl.utils.Reorient2Std(),
-            inputs={
-                'in_file': ('series', nifti_gz_format)},
-            requirements=[fsl_req.v('5.0.9')])
-
-        reorient_epi_opposite = pipeline.add(
-            'reorient_epi_opposite',
-            fsl.utils.Reorient2Std(),
-            inputs={
-                'in_file': ('reverse_phase', nifti_gz_format)},
-            requirements=[fsl_req.v('5.0.9')])
-
-        prep_dwi = pipeline.add(
-            'prepare_dwi',
-            PrepareDWI(
-                topup=True),
-            inputs={
-                'pe_dir': ('ped', str),
-                'ped_polarity': ('pe_angle', str),
-                'dwi': (reorient_epi_in, 'out_file'),
-                'dwi1': (reorient_epi_opposite, 'out_file')})
-
-        ped = pipeline.add(
-            'gen_config',
-            GenTopupConfigFiles(),
-            inputs={
-                'ped': (prep_dwi, 'pe')})
-
-        merge_outputs = pipeline.add(
-            'merge_files',
-            merge_lists(2),
-            inputs={
-                'in1': (prep_dwi, 'main'),
-                'in2': (prep_dwi, 'secondary')})
-
-        merge = pipeline.add(
-            'fsl_merge',
-            fsl_merge(
-                dimension='t',
-                output_type='NIFTI_GZ'),
-            inputs={
-                'in_files': (merge_outputs, 'out')},
-            requirements=[fsl_req.v('5.0.9')])
-
-        topup = pipeline.add(
-            'topup',
-            TOPUP(
-                output_type='NIFTI_GZ'),
-            inputs={
-                'in_file': (merge, 'merged_file'),
-                'encoding_file': (ped, 'config_file')},
-            requirements=[fsl_req.v('5.0.9')])
-
-        in_apply_tp = pipeline.add(
-            'in_apply_tp',
-            merge_lists(1),
-            inputs={
-                'in1': (reorient_epi_in, 'out_file')})
-
-        pipeline.add(
-            'applytopup',
-            ApplyTOPUP(
-                method='jac',
-                in_index=[1],
-                output_type='NIFTI_GZ'),
-            inputs={
-                'in_files': (in_apply_tp, 'out'),
-                'encoding_file': (ped, 'apply_topup_config'),
-                'in_topup_movpar': (topup, 'out_movpar'),
-                'in_topup_fieldcoef': (topup, 'out_fieldcoef')},
-            outputs={
-                'series_preproc': ('out_corrected', nifti_gz_format)},
-            requirements=[fsl_req.v('5.0.9')])
-
-        return pipeline
 
     def _fugue_pipeline(self, **name_maps):
 
@@ -163,7 +63,7 @@ class EpiStudy(MriStudy, metaclass=StudyMetaClass):
             fsl.utils.Reorient2Std(
                 output_type='NIFTI_GZ'),
             inputs={
-                'in_file': ('series', nifti_gz_format)},
+                'in_file': ('magnitude', nifti_gz_format)},
             requirements=[fsl_req.v('5.0.9')])
 
         fm_mag_reorient = pipeline.add(
@@ -215,7 +115,7 @@ class EpiStudy(MriStudy, metaclass=StudyMetaClass):
                 'fmap_in_file': (create_fmap, 'out_fieldmap'),
                 'in_file': (reorient_epi_in, 'out_file')},
             outputs={
-                'series_preproc': ('unwarped_file', nifti_gz_format)},
+                'mag_preproc': ('unwarped_file', nifti_gz_format)},
             wall_time=5,
             requirements=[fsl_req.v('5.0.9')])
 
@@ -249,7 +149,6 @@ class EpiSeriesStudy(EpiStudy, metaclass=StudyMetaClass):
                                "series")),
         InputFilesetSpec('coreg_ref_wmseg', STD_IMAGE_FORMATS,
                          optional=True),
-        InputFilesetSpec('reverse_phase', STD_IMAGE_FORMATS, optional=True),
         InputFilesetSpec('field_map_mag', STD_IMAGE_FORMATS,
                          optional=True),
         InputFilesetSpec('field_map_phase', STD_IMAGE_FORMATS,
@@ -258,6 +157,7 @@ class EpiSeriesStudy(EpiStudy, metaclass=StudyMetaClass):
                     desc=("The magnitude image, typically extracted from "
                           "the provided series")),
         FilesetSpec('series_preproc', nifti_gz_format, 'preprocess_pipeline'),
+        FilesetSpec('mag_preproc', nifti_gz_format, 'mag_preproc_pipeline'),
         FilesetSpec('series_coreg', nifti_gz_format, 'series_coreg_pipeline'),
         FilesetSpec('moco', nifti_gz_format,
                     'intrascan_alignment_pipeline'),
@@ -285,6 +185,36 @@ class EpiSeriesStudy(EpiStudy, metaclass=StudyMetaClass):
         else:
             preproc = 'series_preproc'
         return preproc
+
+    def extract_magnitude_pipeline(self, **name_maps):
+
+        pipeline = self.new_pipeline(
+            'extract_magnitude',
+            desc="Extracts a single magnitude volume from a series",
+            citations=[],
+            name_maps=name_maps)
+
+        pipeline.add(
+            "extract_first_vol",
+            MRConvert(
+                coord=(3, 0)),
+            inputs={
+                'in_file': ('series', nifti_gz_format)},
+            outputs={
+                'magnitude': ('out_file', nifti_gz_format)},
+            requirements=[mrtrix_req.v('3.0rc3')])
+
+        return pipeline
+
+    def preprocess_pipeline(self, **name_maps):
+        return super().preprocess_pipeline(input_map={'magnitude': 'series'},
+                                           name_maps=name_maps)
+
+    def mag_preproc_pipeline(self, **name_maps):
+        self.extract_magnitude_pipeline(
+            input_map={'series': 'series_preproc'},
+            output_map={'magnitude': 'mag_preproc'},
+            name_maps=name_maps)
 
     def coreg_pipeline(self, **name_maps):
         if self.branch('coreg_method', 'epireg'):
@@ -350,26 +280,6 @@ class EpiSeriesStudy(EpiStudy, metaclass=StudyMetaClass):
                 wall_time=10)
         else:
             pipeline = super().brain_coreg_pipeline(**name_maps)
-
-        return pipeline
-
-    def extract_magnitude_pipeline(self, **name_maps):
-
-        pipeline = self.new_pipeline(
-            'extract_magnitude',
-            desc="Extracts a single magnitude volume from a series",
-            citations=[],
-            name_maps=name_maps)
-
-        pipeline.add(
-            "extract_first_vol",
-            MRConvert(
-                coord=(3, 0)),
-            inputs={
-                'in_file': ('series', nifti_gz_format)},
-            outputs={
-                'magnitude': ('out_file', nifti_gz_format)},
-            requirements=[mrtrix_req.v('3.0rc3')])
 
         return pipeline
