+++ test/unittests/study/mri/__init__.py
@@ -1,2 +1,1 @@
-from . import functional
-from . import structural
+
